requirements.txt
# requirements.txt
Flask
Flask-CORS
pytz
psycopg2-binary